Add runshow hdat command
When iterating on an algorithm, I usually want to run a case and then
show the result.  Thus it is useful to have a single command to do this
so that we don't need to string together commands in bash all the time.

In order to make this work as expected, also expanded the resulting
status of running a case.  It can now be:

PASS - there is a golden result, and the current result passes when
compared to it

FAIL - there is a golden result, and the current result fails when
compared to it

UNKNOWN - there is no golden result

ERROR - there was an error running the case